pyproject.toml
[project]
name = "rem_math"
version = "0.1.2"
description = ""
authors = [
    {name = "WrldEngine",email = "kamran_pulatov@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
]


[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "rem_math"
module-name = ".rem_math"

[tool.poetry.group.dev.dependencies]
pytest-benchmark = "^5.1.0"
numpy = "^2.3.1"
maturin = "^1.9.0"
